Home screen feed update
I was fucking up the styles and was struggling with various things so
to avoid some catastrophy I just made a duplicate file for now.
I added “continue watching” section in the feed, made the
feed-item-overlay bigger vertically to create more spacing, minimized
the chevron, made the item-headers darker to create contrast from the
feed-item-overlay, made the names of the concerts and movies in the
feed regular, 1pt larger but same linespacing, letterspacing decreased
to 2.4 from 3.2, color usage minimized.